фиксы киборга (вставка его модулей в постройки)
* Раньше киборг мог вставить свой анализатор атмосферы в камеру и потерять его навсегда. Пофиксил.
* Прошурстев модули киборга, я увидел, что сигналер научного киборга можно так же забаговать и даже хуже. Пофиксил.

+++ code/game/machinery/camera/camera.dm
@@ -154,6 +154,10 @@
 			drop_assembly(1)
 			qdel(src)
 	else if(istype(W, /obj/item/device/analyzer) && panel_open) //XRay
+		if(isrobot(user)){
+			to_chat(user, "<span class='notice'>You cannot insert the [W]-module inside this.</span>")
+			return
+		}
 		if(!isXRay())
 			upgradeXRay()
 			qdel(W)
